virtionet: Move virtio-net code to its own file
Trying to prevent vf/virtio.go from getting too big.
